Change: x, y for GraphPoint and Point back to uint16_t

+++ include/drawobjects/drawobject.h
@@ -68,8 +68,8 @@ typedef struct TextStyle {
 
 typedef struct Point {
 
-    float x;
-    float y;
+    unsigned short int x;
+    unsigned short int y;
 
 } Point;
